+++ Source/Experiments/MJDTestCryoVac/ORMJDTestCryostats/ORMJDTestCryostat.m
@@ -127,7 +127,7 @@ NSString* ORMJDTestCryoConnectionChanged = @"ORMJDTestCryoConnectionChanged";
 {
 	return tag;
 }
-- (void) setTag:(int)aValue
+- (void) setTag:(NSUInteger)aValue
 {
 	tag = aValue;
 }
