+++ Source/Objects/Data Processing/Data Objects/ORDataSetController.m
@@ -93,7 +93,7 @@
 - (void) setUpViews
 {
     int i;
-    int n = [model numberOfChildren];
+    NSUInteger n = [model numberOfChildren];
     for(i=0;i<n;i++){ 
 		ORSubPlotController* subPlotController = [ORSubPlotController panel];
 		[self addSubController:subPlotController];
